workign on refacting the delete post call
