+++ public/assets/js/source/all-posts.js
@@ -97,7 +97,7 @@
 		var $this = $(this);
 
 		var data = {
-			action: 		'process_delete_post',
+			action: 		'process_post_post',
 			postid: 		$this.closest('a').data('postid'),
 			nonce: 			lasso_editor.deletePost
 		}
